+++ src/bunit.web/Rendering/RenderedComponent.cs
@@ -9,12 +9,21 @@ namespace Bunit.Rendering
 	/// <inheritdoc/>
 	internal class RenderedComponent<TComponent> : RenderedFragment, IRenderedComponent<TComponent> where TComponent : IComponent
 	{
+		private readonly TComponent _instance;
+
 		/// <inheritdoc/>
-		public TComponent Instance { get; }
+		public TComponent Instance
+		{
+			get
+			{
+				EnsureComponentNotDisposed();
+				return _instance;
+			}
+		}
 
 		public RenderedComponent(IServiceProvider services, int componentId, TComponent component) : base(services, componentId)
 		{
-			Instance = component;
-		}		
+			_instance = component;
+		}
 	}
 }